demos/requirements.txt
celery==5.3.4
flask==2.3.3
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
numpy==1.25.2
openai==0.28.0
pillow==10.0.1
pycairo==1.24.0
pygobject==3.44.1
redis==5.0.0
